worker/requirements.txt
paho-mqtt==1.6.1
pandas>=2.0
SQLAlchemy>=2.0
python-dateutil>=2.9
python-dotenv>=1.0
psycopg[binary]==3.2.10
